requirements.txt
Flask==2.3.3
Pillow==10.0.0
pytesseract==0.3.10
Werkzeug==2.3.7
gunicorn==21.2.0